Change time.After to time.NewTimer
This fixes a probelm where with 300 replicas on a single node cfluster,
the replicas don't enter 2/2 RUNNING state consistently but instead spam
pilot warnings about port 8080 not being present.

time.After documentation clearly states the following:

"
After waits for the duration to elapse and then sends the
current time on the returned channel. It is equivalent to
NewTimer(d).C. `The underlying Timer is not recovered by the
garbage collector until the timer fires.` If efficiency is a concern,
use NewTimer instead and call Timer.Stop if the timer is no longer needed.
"

In the case of this ads code, we desire to stop the timer before it
expires (or reset it).  This implementation isn't quite right.  Prior
to this PR the timer expired on every single event fed to ADS.
This resulted in sidecars not receiving the appropriate information
they expect.

Note I left two log statements in (The "Done" logs) and they do not
log.  I am not sure why not, other than perhaps the goroutines
aare not exiting naturally and they were always exiting prior because
of the usage of time.After().t

Guidance welcome. :)