Change time.After to time.NewTimer (#13725)
* Change time.After to time.NewTimer

This fixes a problem where with 300 replicas on a single node cfluster,
the replicas don't enter 2/2 RUNNING state consistently but instead spam
pilot warnings about port 8080 not being present until the end of time.

time.After documentation clearly states the following:

"
After waits for the duration to elapse and then sends the
current time on the returned channel. It is equivalent to
NewTimer(d).C. `The underlying Timer is not recovered by the
garbage collector until the timer fires.` If efficiency is a concern,
use NewTimer instead and call Timer.Stop if the timer is no longer needed.
"

In the case of this ads code, we desire to stop the timer before it
expires (or reset it).  This implementation isn't quite right.  Prior
to this PR the timer expired on every single event fed to ADS because
the implementation had a defect.  This resulted in sidecars not
receiving the appropriate information they expect and push timeouts
occuring when no push timeout should occur.

* Address reviewer feedback

* Remove extrenous logging

(cherry picked from commit fdefe93b83c10b016cf77aa140f77a5d0e1e90d5)