Slight refinement - confusing channel logic
- Allocate channels only if the feature flag is set.
- Clean up the alignment a bit

The channel logic in shouldRespond is the minimum set of
channels required to make the software function properly.

Unfortuantely I am not clear why some ACKS match a push and
other ACKS don't match a push.

I have also submitted a related PR https://github.com/istio/istio/pull/27768
that moves the shouldRespond function into the goroutine - so it doesn't
execute in the same thread (main thread?) as the rest of the semaphore
logic. That 3 line PR unfortunately has race conditions as identified
by the test suite, as expected.

it is difficult to tell if additional race conditions are introduced by
this PR or not, will need to fix https://github.com/istio/istio/pull/27768
prior to merging this work.